Remove broken `xdescribe`'d propTypes tests
Fix: removed a test of propType validation that did not work

Neither my solution of spying on console methods
or the solution provided by Keith will work.

1. The spies weren't receiving the messages in all cases.
2. If we test that a propType exists or is equal to a propType
   then it will not work when testing `PropTypes.shape` or
   `PropTypes.arrayOf`.

Since they are so hard to test and are not needing for
the component to function the tests are unimportant and removed.

See: https://github.com/economist-components/component-win-navigation/pull/9#discussion_r41434859